reorganize tests, add new test for `handleCursorChange`
