+++ src/client/routes/~doc/components/AutomergeMonacoBinder/AutomergeMonacoBinder.tsx
@@ -35,6 +35,7 @@ export const AutomergeMonacoBinder = memo(
       }}
       options={{
         automaticLayout: true,
+        cursorStyle: "block",
       }}
     />
   ),
